requirements.txt
dnspython
requests
python-whois
python-Wappalyzer
beautifulsoup4
lxml
